+++ test/codesize/test_codesize_minimal_O0.expected.js
@@ -948,6 +948,7 @@ Module['FS_createPreloadedFile'] = FS.createPreloadedFile;
   'initRandomFill',
   'randomFill',
   'safeSetTimeout',
+  'safeClearTimeout',
   'setImmediateWrapped',
   'safeRequestAnimationFrame',
   'clearImmediateWrapped',